update search script builder

+++ Durchfuehrung/buildsearchscript/src/com/prozess/searchscript/Main.java
@@ -10,26 +10,43 @@ import java.util.List;
 public class Main {
 
     public static void main(String[] args) throws IOException {
-        // write your code here
-        String queryFile = System.getProperty("user.dir") + "\\queries.txt";
-        File file = new File(queryFile);
-        file.delete();
-        BufferedWriter writer = new BufferedWriter(new FileWriter(queryFile , true));
-        for (int i = 0; i < getFirstBlock().size(); i++) {
-            String string1 = getFirstBlock().get(i);
 
-            for ( int j = 0 ; j < getSecondBlock().size(); j++) {
-                String string2 = getSecondBlock().get(j);
-                for ( int k = 0 ; k < getThirdBlock().size(); k++) {
-                    String string3 = getThirdBlock().get(k);
-                    writer.write(string1 + " AND " +  string2 +  " AND " +  string3 +"\n");
+        int querySlection  = 12;
+
+        if (querySlection == 123) {
+            String queryFile = System.getProperty("user.dir") + "\\queries123.txt";
+            File file = new File(queryFile);
+            file.delete();
+            BufferedWriter writer = new BufferedWriter(new FileWriter(queryFile, true));
+            for (int i = 0; i < getFirstBlock().size(); i++) {
+                String string1 = getFirstBlock().get(i);
+
+                for (int j = 0; j < getSecondBlock().size(); j++) {
+                    String string2 = getSecondBlock().get(j);
+                    for (int k = 0; k < getThirdBlock().size(); k++) {
+                        String string3 = getThirdBlock().get(k);
+                        writer.write(string1 + " AND " + string2 + " AND " + string3 + "\n");
+                    }
                 }
             }
-
-
+            writer.close();
         }
 
-        writer.close();
+        if (querySlection == 12) {
+            String queryFile = System.getProperty("user.dir") + "\\queries12.txt";
+            File file = new File(queryFile);
+            file.delete();
+            BufferedWriter writer = new BufferedWriter(new FileWriter(queryFile, true));
+            for (int i = 0; i < getFirstBlock().size(); i++) {
+                String string1 = getFirstBlock().get(i);
+
+                for (int j = 0; j < getSecondBlock().size(); j++) {
+                    String string2 = getSecondBlock().get(j);
+                        writer.write(string1 + " AND " + string2 + "\n");
+                    }
+            }
+            writer.close();
+        }
 
     }
